requirements.txt
albumentations==1.3.0
basicsr==1.4.2
easydict==1.13
einops==0.8.1
imageio==2.9.0
imgaug==0.4.0
joblib==1.4.2
kornia==0.6.0
matplotlib==3.10.1
memory_profiler==0.61.0
numpy==1.24.4
omegaconf==2.1.1
opencv_python==4.9.0.80
opencv_python_headless==4.5.5.64
pandas==2.2.3
Pillow==11.1.0
ptflops==0.7.4
pytorch_lightning==1.5.0
pytorch_msssim==1.0.0
PyYAML==6.0.2
PyYAML==6.0.2
scikit_learn==1.6.1
scipy==1.9.1
six==1.17.0
skimage==0.0
timm==0.4.12
torch==1.12.1
torchvision==0.13.1
tqdm==4.67.1
webdataset==0.2.5
